+++ nym-vpn-android/buildSrc/src/main/kotlin/Constants.kt
@@ -1,8 +1,8 @@
 import org.gradle.api.JavaVersion
 
 object Constants {
-	const val VERSION_NAME = "v3.1.1"
-    const val VERSION_CODE = 31100
+	const val VERSION_NAME = "v3.1.2"
+    const val VERSION_CODE = 31200
     const val TARGET_SDK = 36
     const val COMPILE_SDK = 36
     const val MIN_SDK = 24