+++ nym-vpn-android/CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [3.1.2] - 21.03.2026
 
 ### Changed
-- Use webkpi instead of rustls
+- Use webkpi instead of ruslts-platform-verifier and android only
 
 ## [3.1.1] - 19.03.2026